Plot distributions

Plot the distributions of cells across the tissue with overlaying color information. This is a wrapper function for plot

plotDistributions(
  datIn,
  group,
  groups = NULL,
  colors = rep("black", dim(datIn$mapDat)[2]),
  colormap = gray.colors,
  maxrow = 12,
  pch = 19,
  cex = 1.5,
  xlim = NULL,
  ylim = NULL,
  main = "",
  xlab = "",
  ylab = "",
  singlePlot = TRUE,
  ...
)

Arguments

datIn

a fishScaleAndMap output list

group

a character vector (or factor) indicating how to split the data (e.g., cluster call) or a metadata/mappingResults column name

groups

a character vector of groups to show (default is levels of group)

colors

a character vector (or factor) indicating how to color the plots (e.g., layer or gene expression) or a metadata/mappingResults column name (default is all black)

colormap

function to use for the colormap for the data (default gray.colors)

maxrow

maximum number of plots to show in one row (default=12)

pch, cex

for plot. Can be single values or vectors

xlim, ylim

for plot, but will be calculated if not entered

main, xlab, ylab, ...

other parameters for plot (must be single values)

singlePlot

should everything be plot on a single page (default=TRUE)

Value

Only returns if there is an error