Filter possible cluster calls using priors

This function will return a vector of possible clusters for cells that meet a set of priors for each layer

possibleClustersByPriors(
  cluster,
  layer,
  subsetVector = rep(TRUE, length(cluster)),
  useClusters = sort(unique(cluster)),
  rareLimit = 0.005,
  layerNm = c("L1", "L2/3", "L4", "L5", "L6"),
  scaleByLayer = TRUE,
  scaleByFn = max,
  smartWeight = TRUE,
  spillFactor = 0.15,
  weightCutoff = 0.02
)

Arguments

cluster

vector of all clusters

layer

list of layers for each cluster entry (for data sets with only laminar dissections, each list entry will be of length 1)

subsetVector

a vector of TRUE/FALSE values indicated whether the entry is in the subset of interest (e.g., Cre lines); default is all

useClusters

a set of clusters to be considered a priori (e.g., GABA vs. glut); default is all

rareLimit

define any values less than this as 0. The idea is to exclude rare cells

layerNm

names of all layers. set to NULL to have this calculated

scaleByLayer

if TRUE, scales to the proportion of cells in each layer

scaleByFn

what function should be used for the layer scaling (default=max, ignored if scaleByLayer=FALSE)

smartWeight

if TRUE, multilayer dissections are weighted smartly by cluster, rather than evenly by cluster (FALSE)

spillFactor

fractional amount of cells in a layer below which it is assumed no cells are from that layer in multilayer dissection

weightCutoff

anything less than this is set to 0 for convenience

Value

a vector of possible clusters for cells that meet a set of priors for each layer